requirements.txt
# Django
Django>=6.0.1,<6.1

# Production server
gunicorn>=21.2.0

# Database
psycopg2-binary>=2.9.9  # PostgreSQL adapter
dj-database-url>=2.1.0  # Database URL parsing

# Static files
whitenoise>=6.6.0  # Static file serving

# Image processing (for your image_processing.py)
Pillow>=10.2.0

# Optional but recommended
python-dotenv>=1.0.0  # Environment variable management
